🐛 fix: fix market locale problem
