+++ src/app/api/market/AgentMarket.ts
@@ -1,8 +1,8 @@
 import urlJoin from 'url-join';
 
 import { getServerConfig } from '@/config/server';
-import { DEFAULT_LANG, checkLang } from '@/const/locale';
-import { Locales } from '@/locales/resources';
+import { DEFAULT_LANG, isLocaleNotSupport } from '@/const/locale';
+import { Locales, normalizeLocale } from '@/locales/resources';
 
 export class AgentMarket {
   private readonly baseUrl: string;
@@ -12,14 +12,14 @@ export class AgentMarket {
   }
 
   getAgentIndexUrl = (lang: Locales = DEFAULT_LANG) => {
-    if (checkLang(lang)) return this.baseUrl;
+    if (isLocaleNotSupport(lang)) return this.baseUrl;
 
-    return urlJoin(this.baseUrl, `index.${lang}.json`);
+    return urlJoin(this.baseUrl, `index.${normalizeLocale(lang)}.json`);
   };
 
   getAgentUrl = (identifier: string, lang: Locales = DEFAULT_LANG) => {
-    if (checkLang(lang)) return urlJoin(this.baseUrl, `${identifier}.json`);
+    if (isLocaleNotSupport(lang)) return urlJoin(this.baseUrl, `${identifier}.json`);
 
-    return urlJoin(this.baseUrl, `${identifier}.${lang}.json`);
+    return urlJoin(this.baseUrl, `${identifier}.${normalizeLocale(lang)}.json`);
   };
 }
