+++ src/app/api/plugin/store/Store.ts
@@ -1,7 +1,7 @@
 import urlJoin from 'url-join';
 
 import { getServerConfig } from '@/config/server';
-import { DEFAULT_LANG, checkLang } from '@/const/locale';
+import { DEFAULT_LANG, isLocaleNotSupport } from '@/const/locale';
 import { Locales } from '@/locales/resources';
 
 export class PluginStore {
@@ -12,7 +12,7 @@ export class PluginStore {
   }
 
   getPluginIndexUrl = (lang: Locales = DEFAULT_LANG) => {
-    if (checkLang(lang)) return this.baseUrl;
+    if (isLocaleNotSupport(lang)) return this.baseUrl;
 
     return urlJoin(this.baseUrl, `index.${lang}.json`);
   };
