+++ src/const/locale.ts
@@ -1,9 +1,12 @@
-import { localeOptions } from '@/locales/options';
-import { Locales } from '@/locales/resources';
+import { Locales, supportLocales } from '@/locales/resources';
 
 export const DEFAULT_LANG = 'en-US';
 export const LOBE_LOCALE_COOKIE = 'LOBE_LOCALE';
 
-export const checkLang = (lang: Locales) => {
-  return lang === DEFAULT_LANG || !localeOptions.map((o) => o.value).includes(lang);
+/**
+ * Check if the language is supported
+ * @param locale
+ */
+export const isLocaleNotSupport = (locale: Locales) => {
+  return locale === DEFAULT_LANG || !supportLocales.includes(locale);
 };
