+++ src/features/Conversation/ChatList/Actions/customAction.ts
@@ -2,7 +2,7 @@ import { ActionIconGroupItems } from '@lobehub/ui/es/ActionIconGroup';
 import { LanguagesIcon, Play } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 
-import { localeOptions } from '@/locales/options';
+import { localeOptions } from '@/locales/resources';
 
 export const useCustomActions = () => {
   const { t } = useTranslation('chat');
